Add Docker build
Changed browserlist config to prevent the "productionizer" from rewriting a piece of code in the library '@noble/ed25519' (see 'https://github.com/paulmillr/noble-ed25519/issues/23') which is a transitive dependency of the Concordium libraries.

The library has been updated to not perform the "problematic" operation, but because the Concordium libraries are shipped pre-built, we cannot override the dependency that it uses.