Migrate to concordium-dapp-libraries  (#3)
Replace adhoc implementations of Browser Wallet/WalletConnect integration with the published libraries.

This mostly deletes code related to connection handling that the dapp library now takes care of in a component and a couple of hooks.

Care has been taken to ensure that the migration introduces as few functional changes as possible. In particular, it keeps the behavior that inactive connectors stay alive by using persistentConnectorType. This serves to illustrate that this functionality basically comes for free. Periodic pinging of WalletConnect is included as well.

A few "unrelated" migrations of SDK types are also included.